Move generic venv helpers to own module.
